fix: rename card file to uppercase
